Move the Every macro out of a colliding gigantism.internal object
The `every` collection macro lived in a new `gigantism.internal` object. Because
a wildcard `import gigantism.*` shadows a package member, this hid every
consumer module's own `internal` object (typonym, telekinesis, wisteria, and
~50 others), breaking references like `gigantism.internal.reify` — visible only
in a clean build. Move the macro into `Every.summonAll` in the companion, so
`gigantism` exposes no top-level `internal`.
